+++ Common/src/main/java/at/petrak/hexcasting/api/casting/eval/CastingEnvironment.java
@@ -9,9 +9,11 @@ import at.petrak.hexcasting.api.casting.mishaps.MishapBadLocation;
 import at.petrak.hexcasting.api.casting.mishaps.MishapDisallowedSpell;
 import at.petrak.hexcasting.api.casting.mishaps.MishapEntityTooFarAway;
 import at.petrak.hexcasting.api.mod.HexConfig;
+import at.petrak.hexcasting.api.mod.HexTags;
 import at.petrak.hexcasting.api.pigment.FrozenPigment;
 import at.petrak.hexcasting.api.utils.HexUtils;
 import at.petrak.hexcasting.common.lib.HexAttributes;
+import at.petrak.hexcasting.xplat.IXplatAbstractions;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
@@ -38,6 +40,7 @@ import java.util.function.Predicate;
 
 import static at.petrak.hexcasting.api.HexAPI.modLoc;
 import static at.petrak.hexcasting.api.casting.eval.CastingEnvironmentComponent.*;
+import static at.petrak.hexcasting.api.utils.HexUtils.isOfTag;
 
 /**
  * Environment within which hexes are cast.
@@ -193,14 +196,19 @@ public abstract class CastingEnvironment {
             throw new MishapDisallowedSpell("disallowed", loc);
         }
 
-        costModifier = this.getCostModifier(match);
+        costModifier = (loc != null) ? this.getCostModifier(loc) : 1.0;
     }
 
     /**
-     * Casting env subclasses can override this to modify the cost for a given action
+     * Gets the cost modifier for a given action. By default, this is based on the cost scaling values
+     * in the config. Casting env subclasses can override this to modify the cost in other ways.
      */
-    protected double getCostModifier(PatternShapeMatch match) {
-        return 1.0;
+    protected double getCostModifier(@NotNull ResourceLocation loc) {
+        if (isOfTag(IXplatAbstractions.INSTANCE.getActionRegistry(), loc, HexTags.Actions.CANNOT_MODIFY_COST)) {
+            // blacklisted actions can still be manually scaled in the config, but the global scaling doesn't apply
+            return HexConfig.server().getActionCostScaling(loc);
+        }
+        return HexConfig.server().getActionCostScaling(loc) * HexConfig.server().globalCostScaling();
     }
 
     @Nullable